\name{chart}
\docType{data}
\alias{chart}
\title{Synthetic Control Charts}
\description{
  This data set contains simulated control chart time series of 6 types. See Alcock and Manolopoulos (1999) for details.
}
\usage{chart}
\format{A list with entries $x (a 600x60 matrix with each row corresponding to an individual time series) and $c (a vector of labels indicating the chart type).}
\source{UCI Machine Learning Repository.}
\references{
  Alcock R.J. and Manolopoulos Y. (1999) Time-Series Similarity Queries Employing a Feature-Based Approach. 7th Hellenic Conference on Informatics.
}
\keyword{datasets}
