% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_Binary.R
\name{marker_Binary}
\alias{marker_Binary}
\title{Title}
\usage{
marker_Binary(bulk_data, features, ref_group, Log2FC_cutoff = 0.585)
}
\arguments{
\item{bulk_data}{Log2-normalized bulk expression data with genes in row and samples in column.}

\item{features}{Feature data of bulk samples, column1 are sample names and column2 are feature labels of each sample.}

\item{ref_group}{A character to indicate which feature is the control group.}

\item{Log2FC_cutoff}{Absolute cutoff value of fold change, default is 0.585.}
}
\value{
A gene list of feature markers with two binary groups.
}
\description{
Title
}
