% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_Continuous.R
\name{marker_Continuous}
\alias{marker_Continuous}
\title{Title}
\usage{
marker_Continuous(bulk_data, features, estimate_cutoff = 0.2)
}
\arguments{
\item{bulk_data}{Log2-normalized bulk expression data with genes in row and samples in column.}

\item{features}{Feature data of bulk samples, such as TMB or CNA values of each sample.}

\item{estimate_cutoff}{Absolute cutoff value of correlation coefficient, default is 0.2.}
}
\value{
A gene list of feature markers.
}
\description{
Title
}
