% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_Survival.R
\name{marker_Survival}
\alias{marker_Survival}
\title{Title}
\usage{
marker_Survival(bulk_data, survival_data)
}
\arguments{
\item{bulk_data}{Log2-normalized bulk expression data with genes in row and samples in column.}

\item{survival_data}{Survival data with time in column1 and status in column2. Rownames are sample name.}
}
\value{
A gene list.
}
\description{
Title
}
