% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_a.R
\name{optimal_a}
\alias{optimal_a}
\title{Title}
\usage{
optimal_a(sc_dataset, geneList, probs = seq(0.2, 0.45, by = 0.05))
}
\arguments{
\item{sc_dataset}{A seurat object of single cell RNA sequencing data.}

\item{geneList}{A gene list correlated with interested features.}

\item{probs}{Cutoff value of ScPP, default is seq(0.2, 0.45, by = 0.05).}
}
\value{
The optimal prob based on NES.
}
\description{
Title
}
