% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_ScPP.R
\name{OptimizationMode}
\alias{OptimizationMode}
\title{Optimization Mode for Single-Cell Perturbation Prediction}
\usage{
OptimizationMode(
  sc_dataset,
  geneList,
  probs,
  verbose = SigBridgeRUtils::getFuncOption("verbose") \%||\% TRUE,
  parallel = !inherits(future::plan("list")[[1]], "sequential"),
  seed = SigBridgeRUtils::getFuncOption("seed") \%||\% 123L,
  assay = assay,
  ...
)
}
\arguments{
\item{sc_dataset}{Single-cell dataset (Seurat object)}

\item{geneList}{List containing gene sets for AUC calculation and GSEA}

\item{probs}{Vector of probability thresholds to test}

\item{verbose}{Whether to show progress messages}

\item{parallel}{Whether to use parallel processing}

\item{seed}{Random seed for reproducibility}

\item{assay}{Slot name of assay to be used}

\item{...}{For future updates}
}
\value{
Optimal probability threshold that maximizes NES difference
}
\description{
Identifies optimal probability threshold for cell classification by testing
multiple thresholds and selecting the one that maximizes normalized enrichment
score (NES) difference between high and low perturbation groups.
}
\seealso{
\code{\link[=ScPP.optimized]{ScPP.optimized()}}

Other scPP: 
\code{\link{FixedProbMode}()},
\code{\link{ScPP.optimized}()},
\code{\link{marker_Binary.optimized}()},
\code{\link{marker_Continuous.optimized}()},
\code{\link{marker_Survival2}()}

Other scPP_optimal_param: 
\code{\link{ProcessSingleProb}()}
}
\concept{scPP}
\concept{scPP_optimal_param}
