% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singcell_Preprocessing.R
\name{sc_Preprocess}
\alias{sc_Preprocess}
\title{Title}
\usage{
sc_Preprocess(
  counts,
  project = "sc_preprocess",
  normalization.method = "LogNormalize",
  scale.factor = 10000,
  selection.method = "vst",
  resolution = 0.1,
  dims_Neighbors = 1:20,
  dims_TSNE = 1:20,
  dims_UMAP = 1:20
)
}
\arguments{
\item{counts}{Count data of scRNA sequencing.}

\item{project}{Project name.}

\item{normalization.method}{Normalization method.}

\item{scale.factor}{}

\item{selection.method}{}

\item{resolution}{}

\item{dims_Neighbors}{}

\item{dims_TSNE}{}

\item{dims_UMAP}{}
}
\value{
A seurat object.
}
\description{
Title
}
