\name{SCPP-package}
\alias{SCPP-package}
\alias{SCPP}
\docType{package}
\title{
\packageTitle{SCPP}
}
\description{
Performs projection pursuit based on spectral connectivity to find optimal subspaces
    for spectral clustering.
    Clustering is performed recursively by repeatedly identifying subspaces for separating (some)
    clusters from others. All datasets included are from the UCI machine learning repository 
[https://archive.ics.uci.edu/ml/datasets.html](Bache and Lichman, 2013)
}
\author{
David Hofmeyr[aut, cre]
Maintainer: \packageMaintainer{PPCI}
}
\references{
Hofmeyr, D., Pavlidis, N., Eckley, I. (2018) Minimum spectral connectivity projection pursuit, \emph{Statistics and Computing}, to appear.
}
\keyword{ package }
