% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScPP.R
\name{ScPP}
\alias{ScPP}
\title{Title}
\usage{
ScPP(sc_dataset, geneList, probs = 0.2)
}
\arguments{
\item{sc_dataset}{A seurat object of single cell RNA sequencing data.}

\item{geneList}{A gene list correlated with interested features.}

\item{probs}{Cutoff value of ScPP, default is 0.2.}
}
\value{
Meta data with information of ScPP selected cells.
}
\description{
Title
}
