\name{SCPP_plot}
\alias{SCPP_plot}
\title{Visualise a Clustering Model Generated by SCPP}
\description{
  Provides a visualisation of a hierarchical clustering model via two-dimensional projections of the data assigned to each node.
}
\usage{
SCPP_plot(sol, labels, node.numbers)
}
\arguments{
  \item{sol}{a clustering solution produced by the function SCPP_cluster.}
  \item{labels}{(optional) a vector of class labels. if provided then points in different classes are plotted in different colours.}
  \item{node.numbers}{(optional) logical. if TRUE then numbers are added to the plot to indicate the order in which nodes were added to the model. if omitted then node.numbers = TRUE.}
}
\keyword{file}
